requirements.txt
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
charset-normalizer==2.1.1
colorama==0.4.6
coloring==0.1.8
colorlog==6.7.0
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.5
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
threadpoolctl==3.1.0
torch==1.13.0+cu116
torch-cluster==1.6.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse==0.6.15
torch-spline-conv==1.2.1
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
wincertstore==0.2
